+++ lib/puppeteer_wrapper/puppeteer_types.ts
@@ -6,5 +6,5 @@ export {
     Page, Frame, Viewport, EvaluateFn, SerializableOrJSHandle, JSHandle, Response, Worker,
     ScriptTagOptions, Browser, Base64ScreenShotOptions, Keyboard, Mouse, NavigationOptions, WaitForSelectorOptions, ElementHandle,
     Touchscreen, Cookie, SetCookie, DeleteCookie, PageEventObj, Request, Timeoutable, PDFOptions, ConsoleMessage, ConsoleMessageType,
-    ResourceType, DialogType, Dialog, BrowserContext, Target, BrowserContextEventObj, GeoOptions, Permission
+    ResourceType, DialogType, Dialog, BrowserContext, Target, BrowserContextEventObj, GeoOptions, Permission, MediaType, MediaFeature
 } from 'puppeteer';